+++ grafana/build-plugin.sh
@@ -4,5 +4,5 @@ yarn install --pure-lockfile
 yarn build
 mage -v
 mage build:generateManifestFile
-cp -r dist micromegas-datasource
-zip -r micromegas-datasource.zip micromegas-datasource
+cp -r dist micromegas-micromegas-datasource
+zip -r micromegas-micromegas-datasource.zip micromegas-micromegas-datasource